+++ src/main/java/com/account/yomankum/web/response/Response.java
@@ -1,4 +1,4 @@
-package com.account.yomankum.web;
+package com.account.yomankum.web.response;
 
 import lombok.Builder;
 import lombok.Getter;
